+++ walkers/src/style/mod.rs
@@ -107,6 +107,7 @@ pub enum Layer {
     #[serde(rename_all = "kebab-case")]
     Symbol {
         source_layer: SourceLayer,
+        minzoom: Option<f32>,
         filter: Option<Filter>,
         layout: Layout,
         paint: Option<Paint>,