+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 * `walkers::Geometry` is replaced by the re-exported `walkers::geo_types`.
+* `symbol` layers support `minzoom`.
 
 ## 0.60.0